Added unit tests

+++ routes/download.js
@@ -6,9 +6,13 @@ var fs = require('fs');
 
 router.get('/', function(req, res) {
   res.download('./output/StatusGram.pdf');
-  res.on('finish', function() {
-    fs.unlinkSync('./output/StatusGram.pdf');
-  });
+
+  if (process.env.COMPUTERNAME !== 'A004651') {
+    res.on('finish', function() {
+      fs.unlinkSync('./output/StatusGram.pdf');
+    });
+  }
+  
 });
 
 module.exports = router;
